+++ src/redux/user/index.ts
@@ -1,0 +1,2 @@
+export { getUser, search } from './action';
+export { selectUser } from './selector';