+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+# [0.2.0](https://github.com/Howard86/github-search/compare/v0.1.0...v0.2.0) (2021-01-22)
+
+
+### Bug Fixes
+
+* unfound username from github ([f745dbb](https://github.com/Howard86/github-search/commit/f745dbbb414bdae120ba4e853995724cbb8b8c00))
+* update UserCard styles ([0f37736](https://github.com/Howard86/github-search/commit/0f3773671d2f4a84c2e1df79e4cec8a59c889b86))
+* UserCard test ([959245a](https://github.com/Howard86/github-search/commit/959245aa03f0617c8e510c49aec6e79fcf9dfb6b))
+
+
+### Features
+
+* add custom progress Loader ([5f2e42d](https://github.com/Howard86/github-search/commit/5f2e42dafba7ecbdb47bbadd5ea6f6e1867c09ca))
+* add GET /users & /users/[username] API ([5a5fbbb](https://github.com/Howard86/github-search/commit/5a5fbbb817ab54ef8b814b30c60002f45bbcab9f))
+* add github user service ([3a2ba2c](https://github.com/Howard86/github-search/commit/3a2ba2ca0bba2cd2487540e5ba7001802fa2a229))
+* add pagination from user API ([f6d55f7](https://github.com/Howard86/github-search/commit/f6d55f7c430aeed39fa8aa62ba72759b0eac5159))
+* add SearchBar ([f3f4916](https://github.com/Howard86/github-search/commit/f3f491681ba59ec6c837d6c8d30a24966f680e77))
+* add user repos, followers & followings ([13b0482](https://github.com/Howard86/github-search/commit/13b0482e25f0ba7a3f04ab8ff88f7ca0e09a3446))
+* add user store for search and single get ([0e7ac59](https://github.com/Howard86/github-search/commit/0e7ac59b12cc372d48c5a003662be25e7a81961a))
+* add UserCard ([c56b71c](https://github.com/Howard86/github-search/commit/c56b71cdabb60671ec60cbe80cfe1cfb65054f80))
+* add UserPage with dynamic username ([b31fd66](https://github.com/Howard86/github-search/commit/b31fd66080505b16683466022722c4e3197b3fa0))
+* update HomePage ([34365f0](https://github.com/Howard86/github-search/commit/34365f05a788556b1d83795a759ecc7ee31c931b))
+* update SearchBar with PaginationButton ([33223e7](https://github.com/Howard86/github-search/commit/33223e76ee7df05b720b2aa9029e71480c6857ff))
+
+
+
 # 0.1.0 (2021-01-08)